notification_service/requirements.txt
fastapi
uvicorn[standard]
pymongo
pydantic[email] # Though email might not be directly validated here, good practice
python-dotenv
aio_pika # Asynchronous RabbitMQ client